Suppression de la grille intermédiaire inutile
